Add preamble collection records
I originally was attempting to take the entire preamble and store that,
but it turned into a nightmare, so for now just taking the most
important parts will have to do.

The 39 character limit did not seem to be able to be overcome by using
the waveform record. It also did not seem to want to take a semicolon
separator anyway.